requirements.txt
# Core ML Libraries (PyTorch는 install_vllm.sh에서 환경별로 설치)
# torch>=2.4.0  # 스크립트에서 환경별 설치
# torchvision>=0.19.0  # 스크립트에서 환경별 설치
# git+https://github.com/huggingface/transformers.git  # 스크립트에서 별도 설치
transformers>=4.43.0
accelerate>=0.25.0
optimum>=1.14.0
# auto-gptq>=0.7.0  # Apple Silicon에서 문제 가능성, 필요시 수동 설치
sentence-transformers>=3.1.0
einops>=0.7.0
# numpy<2.0  # jinaai/jina-embeddings-v3에서 요구

# VLM & Image Processing  
Pillow>=10.0.0
opencv-python>=4.8.0
# paddlepaddle>=2.5.0  # Apple Silicon에서 설치 문제 가능성
# paddleocr>=2.7.0     # Apple Silicon에서 설치 문제 가능성
# trocr>=0.1.0         # 패키지 없음, 제거
qwen-vl-utils>=0.0.10
av>=10.0.0

# High-Performance Inference with vLLM (스크립트에서 환경별로 설치)
# vllm>=0.7.2  # 스크립트에서 환경별 설치
# flash-attn>=2.6.1  # Optional: for memory optimization
# xformers>=0.0.22  # Optional: for additional optimizations

# vLLM 관련 추가 의존성
ray>=2.8.0  # vLLM의 분산 처리
ninja>=1.11.1  # 컴파일 최적화
packaging>=21.0

# RAG & Vector DB
langchain>=0.3.0
langgraph>=0.2.0
langsmith>=0.2.0  # LangSmith tracing and observability
chromadb>=0.4.0
pgvector>=0.2.0
psycopg2-binary>=2.9.7  # Better compatibility on Linux
faiss-cpu>=1.7.4
# Web Search & Knowledge Enhancement
tavily-python>=0.3.0

# PDF Processing
PyMuPDF>=1.23.0  # fitz is included in PyMuPDF
pdf2image>=1.16.0
unstructured[pdf]>=0.15.0  # UnstructuredPDFLoader 지원
unstructured-client>=0.25.0  # Unstructured API 클라이언트 (선택사항)

# CAD/DWG Processing
ezdxf>=1.3.0  # DXF/DWG 파일 읽기 및 분석

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
shapely>=2.0.0  # Geometric operations
matplotlib>=3.7.0  # Plotting
seaborn>=0.12.0  # Statistical visualization

# CUDA support (if available)
# torch-audio
# torchtext

# Web UI & Dashboard
solara>=1.35.0
ipywidgets>=8.1.0

# Development
jupyterlab>=4.0.0

chromadb
networkx>=3.0 # 그래프 자료구조 및 알고리즘

# LangChain/LangGraph packages
langchain>=0.1.0
langchain-core>=0.1.0  
langchain-community>=0.0.20
langgraph>=0.0.26
langsmith>=0.0.80

# LangChain/LangGraph packages
